+++ MailScanner_perl_scripts/MailWatchConf.pm
@@ -16,8 +16,8 @@ my ($db_pass) = 'mailwatch';
 # Change the value below for SQLSpamSettings.pm (default = 15)
 my ($ss_refresh_time) = 15;       # Time in minutes before lists are refreshed
 
-# Change the value below for SQLBlackWhiteList.pm (default = 15)
-my ($bwl_refresh_time) = 15;      # Time in minutes before lists are refreshed
+# Change the value below for SQLAllowBlockList.pm (default = 15)
+my ($abl_refresh_time) = 15;      # Time in minutes before lists are refreshed
 
 
 ###############################
@@ -28,7 +28,7 @@ sub mailwatch_get_db_name { return $db_name };
 sub mailwatch_get_db_host { return $db_host };
 sub mailwatch_get_db_user { return $db_user };
 sub mailwatch_get_db_password { return $db_pass };
-sub mailwatch_get_BWL_refresh_time { return $bwl_refresh_time };
+sub mailwatch_get_ABL_refresh_time { return $abl_refresh_time };
 sub mailwatch_get_SS_refresh_time { return $ss_refresh_time };
 
 1;